FizzBuzz upload
This program uses conosle.log to print all the numbers from 1 to 100.

If any of those numbers are divisible by 3, print Fizz instead of the number.

If any of those numbers are divisible by 5, print Buzz instead of the number.

If the number is divisible by both, print FizzBuzz instead of the number.