initial upload
This algorithm involves two functions.

Range takes integers and returns an array of numbers within a set range between and inclusive of start and finish. E.G argument 1, 10 will return 1,2,3...10

sumOfArray simply adds up all the numbers in an array.

In the example, I've included range as the argument for sumOfArray to add all of the numbers in a range.